create SendMessageEvent
