+++ lightyear/src/server/events.rs
@@ -277,7 +277,7 @@ pub type ComponentRemoveEvent<C> =
     crate::shared::events::components::ComponentRemoveEvent<C, ClientId>;
 
 /// Bevy [`Event`] emitted on the server on the frame where a (non-replication) message is received
-pub type MessageEvent<M> = crate::shared::events::components::MessageEvent<M>;
+pub type MessageEvent<M> = crate::shared::events::message::MessageEvent<M>;
 
 #[cfg(test)]
 mod tests {
